Misc fixes

+++ instructor/introduction/org/jetbrains/kotlinworkshop/introduction/_2ControlStructures/_1IfAndWhen.kt
@@ -6,7 +6,7 @@ enum class Color {
 }
 
 fun updateWeather(
-        celsiusDegrees: Double
+        celsiusDegrees: Int
 ) {
     val description: String
     val color: Color
